+++ src/babel/__tests__/examples/stylesheet_colors.js
@@ -1,0 +1,24 @@
+/* eslint-disable */
+import zacs from '@nozbe/zacs'
+
+const styles = zacs.stylesheet({
+  standardColors: {
+    backgroundColor: '#f0f',
+    borderBottomColor: '#ff00ff',
+    borderLeftColor: '#ff00ff00',
+    borderRightColor: 'rgb(255, 0, 255)',
+    color: 'rgba(255, 0, 255, 1.0)',
+    borderColor: 'burlywood',
+  },
+  unusualColors: {
+    borderColor: '#f0ff',
+    borderTopColor: 'rgb(255 0 255)',
+    overlayColor: 'rgba(255 0 255 / 1.0)',
+    shadowColor: 'hsl(360, 100%, 100%)',
+    textDecorationColor: 'hsl(360 100% 100%)',
+    textShadowColor: 'hsla(360, 100%, 100%, 1.0)',
+    tintColor: 'hsla(360 100% 100% / 1.0)',
+    backgroundColor: 'hwb(360, 100%, 100%)',
+    borderBottomColor: 0xff00ff00,
+  },
+})